Add RPM-like packaging
This commit adds basic scripts that are easily runnable on any pc to
generate and test rpm images, and also adds a github workflow that
creates RPM packages for every commit to the main branch.

The github workflow for now only builds without testing, as we are
getting some strange errors because of docker/podman-in-docker/podman.